HuePi-Backend/requirements.txt
zeroconf~=0.132.2
fastapi~=0.111.1
python-dotenv~=1.0.1
requests~=2.32.3
urllib3~=2.2.2
starlette~=0.37.2
smbus2~=0.4.3
numpy~=2.0.1
rgbxy~=0.5
RPi.bme280~=0.2.4